Obsolete tensor_base.h, refactor everything to tensor.h
